structure runtime tests
