+++ runtime/src/tests/mod.rs
@@ -1,0 +1,1 @@
+mod multiplier;
